d2f

+++ src/blas/matrix/matadd/float_matadd.cpp
@@ -4,8 +4,8 @@
 namespace monolish {
 
 // Dense ///////////////////
-void blas::matadd(const matrix::Dense<double> &A,
-                  const matrix::Dense<double> &B, matrix::Dense<double> &C) {
+void blas::matadd(const matrix::Dense<float> &A,
+                  const matrix::Dense<float> &B, matrix::Dense<float> &C) {
   Logger &logger = Logger::get_instance();
   logger.func_in(monolish_func);
 
@@ -28,8 +28,8 @@ void blas::matadd(const matrix::Dense<double> &A,
 }
 
 // CRS ///////////////////
-void blas::matadd(const matrix::CRS<double> &A, const matrix::CRS<double> &B,
-                  matrix::CRS<double> &C) {
+void blas::matadd(const matrix::CRS<float> &A, const matrix::CRS<float> &B,
+                  matrix::CRS<float> &C) {
   Logger &logger = Logger::get_instance();
   logger.func_in(monolish_func);
 
@@ -62,8 +62,8 @@ matrix::Dense<T> matrix::Dense<T>::operator+(const matrix::Dense<T> &B) {
 
   return C;
 }
-template matrix::Dense<double>
-matrix::Dense<double>::operator+(const matrix::Dense<double> &B);
+template matrix::Dense<float>
+matrix::Dense<float>::operator+(const matrix::Dense<float> &B);
 
 
 template <typename T>
@@ -77,7 +77,7 @@ matrix::CRS<T> matrix::CRS<T>::operator+(const matrix::CRS<T> &B) {
 
   return C;
 }
-template matrix::CRS<double>
-matrix::CRS<double>::operator+(const matrix::CRS<double> &B);
+template matrix::CRS<float>
+matrix::CRS<float>::operator+(const matrix::CRS<float> &B);
 
 } // namespace monolish
